#57 Исправил ошибки, перетащил весь функционал MasterPassword и проч. внутрь Input
