+++ src/pitaya/components/Input/Readme.md
@@ -1,0 +1,23 @@
+# Text Component
+
+Component responsible for inputs.
+
+# Example
+
+```jsx
+<Input
+  value={password}
+  label='Master Password'
+  type='password'
+  onChange={(value) => {changePassword(value);}}
+/>
+```
+
+```jsx
+<Input
+  value={domainName}
+  label='URL'
+  onChange={(value) => {changeDomainName(value);}}
+  buttonOnClick={() => {console.log('CLICK!');}}
+/>
+```
